Scheduler/requirements.txt
kafka-python
sockets
schedule 
flask
requests
python-dotenv